+++ libraries/test_session_current/src/androidTest/java/androidx/media3/session/MediaControllerTest.java
@@ -2353,6 +2353,24 @@ public class MediaControllerTest {
   }
 
   @Test
+  public void getUnmuteVolume_returnsUnmuteVolumeOfPlayerInSession_roundTrip() throws Exception {
+    float testVolume = .5f;
+
+    Bundle playerConfig =
+        new RemoteMediaSession.MockPlayerConfigBuilder().setVolume(testVolume).build();
+    remoteSession.setPlayer(playerConfig);
+
+    MediaController controller = controllerTestRule.createController(remoteSession.getToken());
+    threadTestRule.getHandler().postAndSync(controller::mute);
+    float volume = threadTestRule.getHandler().postAndSync(controller::getVolume);
+    assertThat(volume).isEqualTo(0);
+
+    threadTestRule.getHandler().postAndSync(controller::unmute);
+    volume = threadTestRule.getHandler().postAndSync(controller::getVolume);
+    assertThat(volume).isEqualTo(testVolume);
+  }
+
+  @Test
   public void getCurrentMediaItemIndex() throws Exception {
     int testMediaItemIndex = 1;
     Bundle playerConfig =